refactor: mindahin hero, navbar, partners
