+++ app/layout.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import "./globals.css";
 import TanStackProvider from "@/providers/tanstackQueryProvider";
-import Navbar from "./(landing_page)/navbar";
+import Navbar from "../components/landing/navbar";
 
 const inter = Inter({ subsets: ["latin"] });
