+++ components/landing/partner.tsx
